requirements.txt
# Requirements for the p20w dashboards websites
# running on Python 3.6.0

django == 2.1.2
psycopg2 >=2.6.2 # if using postgres
psycopg2-binary
django-mathfilters >= 0.4.0
django-tinymce >= 2.7.0
Pillow
gunicorn
django-admin-sortable2